[#959] Refactored methods to common class
