+++ app/src/main/java/org/akvo/flow/util/FormFileUtil.java
@@ -25,12 +25,10 @@ import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
 import java.io.File;
-import java.util.ArrayList;
-import java.util.List;
 
 import javax.inject.Inject;
 
-public class FormFileUtil {
+public class FormFileUtil extends InternalFileUtil{
 
     private static final String DIR_FORMS = "forms";
 
@@ -38,79 +36,15 @@ public class FormFileUtil {
     public FormFileUtil() {
     }
 
+    @Override
     @NonNull
-    public File getFormsFolder(Context context) {
-        String path = FileUtil.getInternalFolderPath(context, DIR_FORMS);
-        return FileUtil.createDir(path);
+    protected String getInternalFolderPath(Context context) {
+        return FileUtil.getInternalFolderPath(context, DIR_FORMS);
     }
 
-    @NonNull
-    public File findFormFile(Context context, String formFileName) {
-        File file = getAppInternalFormFileIfExists(context, formFileName);
-        if (file == null) {
-            file = getAppExternalFormFileIfExists(context, formFileName);
-        }
-        if (file == null) {
-            file = getPublicFormFile(formFileName);
-        }
-        return file;
-    }
-
-    @NonNull
-    public List<File> findAllPossibleFormFolders(Context context) {
-        List<File> folders = new ArrayList<>(3);
-        File folder = getAppInternalFormsDir(context);
-        if (folder.exists()) {
-            folders.add(folder);
-        }
-        File folder2 = getAppExternalFormsDir(context);
-        if (folder2 != null && folder2.exists()) {
-            folders.add(folder2);
-        }
-        File folder3 = getPublicFormsDir();
-        if (folder3.exists()) {
-            folders.add(folder3);
-        }
-        return folders;
-    }
-
-    @NonNull
-    private File getPublicFormFile(String formFileName) {
-        return new File(getPublicFolderPath(), formFileName);
-    }
-
+    @Override
     @Nullable
-    private File getAppInternalFormFileIfExists(Context context, String formFileName) {
-        File folder = getAppInternalFormsDir(context);
-        if (folder.exists()) {
-            File file = new File(getFormsFolder(context), formFileName);
-            if (file.exists()) {
-                return file;
-            }
-        }
-        return null;
-    }
-
-    @NonNull
-    private File getAppInternalFormsDir(Context context) {
-        String path = FileUtil.getInternalFolderPath(context, DIR_FORMS);
-        return new File(path);
-    }
-
-    @Nullable
-    private File getAppExternalFormFileIfExists(Context context, String formFileName) {
-        File folder = getAppExternalFormsDir(context);
-        if (folder != null && folder.exists()) {
-            File file = new File(folder, formFileName);
-            if (file.exists()) {
-                return file;
-            }
-        }
-        return null;
-    }
-
-    @Nullable
-    private File getAppExternalFormsDir(Context context) {
+    protected File getAppExternalFormsDir(Context context) {
         String path = FileUtil.getAppExternalFolderPath(context, DIR_FORMS);
         File folder = null;
         if (path != null) {
@@ -119,15 +53,9 @@ public class FormFileUtil {
         return folder;
     }
 
+    @Override
     @NonNull
-    private File getPublicFormsDir() {
-        return new File(FileUtil.getPublicFolderPath(DIR_FORMS));
+    protected String getPublicFolderPath() {
+        return FileUtil.getPublicFolderPath(DIR_FORMS);
     }
-
-    @NonNull
-    private File getPublicFolderPath() {
-        String publicFolderPath = FileUtil.getPublicFolderPath(DIR_FORMS);
-        return FileUtil.createDir(publicFolderPath);
-    }
-
 }
